Add files via upload
Movietickets:This project calculates the cost of movie tickets based on a person's age. It offers different ticket prices depends upon the age group.
StudentGrade:This project determines a student's final grade based on the average of marks in five subjects. It also indicates whether the student has passed or failed.
Anagram: This project helps users to find anagrams of a given word within a list of words.